refactor: simplefy the style in ContentContact and Footer, remove styled-component
