+++ README.md
@@ -6,11 +6,10 @@ This repository contains the TypeScript codebase of a website for "Jardinieresma
 
 ## Features
 
-- **Achieved scalable architecture:** by TypeScript + React 
-- **Obtains contents always up-to-date:** Assembled with Facebook API & Flickr API. 
-- **Appeal to the eye:** Enhance visuals with react-image-gallery and styled-component. 
-- **Deployment automation:** Integrate Github Actions for smoother development. 
-
+- **Achieved scalable architecture:** by TypeScript + React
+- **Obtains contents always up-to-date:** Assembled with Facebook API & Flickr API.
+- **Appeal to the eye:** Enhance visuals with react-image-gallery.
+- **Deployment automation:** Integrate Github Actions for smoother development.
 
 ## Demo
 
@@ -30,15 +29,16 @@ To view or contribute to the "Jardinieresmasquees" project, follow these steps:
 
 1. Navigate to the project directory:
 
-    ```shell
-    cd Jardinieresmasquees
-    ```
+   ```shell
+   cd Jardinieresmasquees
+   ```
 
-2. Explore the codebase, contribute your own garden designs, or engage with the community.
+1. Explore the codebase, contribute your own garden designs, or engage with the community.
 
-※This project employs a commitlinter; when committing, please make a commit message following  with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
+※This project employs a commitlinter; when committing, please make a commit message following with [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
 ## License
+
 This project is open-source and available under the MIT License. See the LICENSE file for details.
 
 Explore, create, and connect with fellow gardening enthusiasts through "Jardinieresmasquees." Join and let your creativity blossom!
