+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "vite",
+    "dev:0.0.0.0": "vite --host 0.0.0.0",
     "build": "tsc && vite build",
     "preview": "vite preview",
     "start": "npm run build && serve -s build --no-clipboard"
@@ -23,7 +24,6 @@
     "react-dom": "^19.2.4",
     "react-image-gallery": "^2.1.2",
     "react-router-dom": "^7.13.1",
-    "styled-components": "^6.0.7",
     "tailwindcss": "^4.2.1"
   },
   "devDependencies": {
