+++ src/component/footer.component.tsx
@@ -1,0 +1,10 @@
+const Footer = () => {
+  const groupName = "Jardinieres masquées de Tours";
+  return (
+    <footer>
+      <p className="text-center md:text-xl">&#169;{groupName}</p>
+    </footer>
+  );
+};
+
+export default Footer;
