--- src/component/footer/footer.component.tsx
@@ -1,14 +1,0 @@
-import { FooterContainer } from "./footer.styles";
-
-const Footer = () => {
-  const groupName = "Jardinieres masquées de Tours";
-  return (
-    <FooterContainer>
-      <div className="footer-sub-container">
-        <div>&#169;{groupName}</div>
-      </div>
-    </FooterContainer>
-  );
-};
-
-export default Footer;
